Addressed exception reported by Matt Mears when entering AirportEats (toString() called on object that may be null).
Added latitude and longitude into the Google Maps query so the map still
goes to the spot if there's a lat + long but no address. The
ActionView method being used to invoke Google Maps is functional, but
behaves unpredictably when the address is not entered in a valid form by
the user who uploaded it. We may want to consider using MapView
depending on what is encountered in further testing.

Added ability for user to be saved to application even if they don't hit
remember me. This won't affect the functionality of remember me, but it
does allow for better handling of situations where memory may be
cleared, and the user needs to retrieved again. I didn't do anything
with this yet, but started to implement this intending to do the same
for city due to exceptions encountered in testing. Exceptions stopped
occurring after other changes though, so may have been indirectly
resolved.