MattG- What I have so far for ImageUpload.java and 2 accompanying xml files (layout and menu).
You can delete the java/xml related to ImageUploader ("er" at the end
of file name) if the push doesn't do it. ImageUpload has compile errors,
so I've commented most of the code out, but a resource conflict with the
xml files also causes an error. Sorry if this hinders anyone else work.